requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
werkzeug
anthropic
byaldi